Cleanup.  Painter's Flower Pot merged.  Small things. v1.9.9 candidate.

+++ carpentersblocks/renderer/helper/ParticleHelper.java
@@ -1,18 +1,14 @@
 package carpentersblocks.renderer.helper;
 
-import carpentersblocks.tileentity.TEBase;
-import carpentersblocks.util.BlockProperties;
-import carpentersblocks.util.handler.EventHandler;
-import carpentersblocks.util.handler.OverlayHandler;
 import net.minecraft.block.Block;
 import net.minecraft.client.particle.EffectRenderer;
 import net.minecraft.client.particle.EntityDiggingFX;
 import net.minecraft.entity.Entity;
-import net.minecraft.item.ItemBlock;
-import net.minecraft.item.ItemStack;
 import net.minecraft.util.MovingObjectPosition;
 import net.minecraft.world.World;
-import cpw.mods.fml.client.FMLClientHandler;
+import carpentersblocks.tileentity.TEBase;
+import carpentersblocks.util.BlockProperties;
+import carpentersblocks.util.handler.OverlayHandler;
 
 public class ParticleHelper {
 
@@ -27,16 +23,16 @@ public class ParticleHelper {
 	public static void spawnTileParticleAt(World world, Entity entity, int blockID, int metadata)
 	{
 		entity.worldObj.spawnParticle(
-			"tilecrack_" + blockID + "_" + metadata,
-			entity.posX + ((double) entity.worldObj.rand.nextFloat() - 0.5D) * (double) entity.width,
-			entity.boundingBox.minY + 0.1D,
-			entity.posZ + ((double) entity.worldObj.rand.nextFloat() - 0.5D) * (double) entity.width,
-			-entity.motionX * 4.0D,
-			1.5D,
-			-entity.motionZ * 4.0D
-		);
+				"tilecrack_" + blockID + "_" + metadata,
+				entity.posX + (entity.worldObj.rand.nextFloat() - 0.5D) * entity.width,
+				entity.boundingBox.minY + 0.1D,
+				entity.posZ + (entity.worldObj.rand.nextFloat() - 0.5D) * entity.width,
+				-entity.motionX * 4.0D,
+				1.5D,
+				-entity.motionZ * 4.0D
+				);
 	}
-	
+
 	/**
 	 * Produces block destruction particles at coordinates.
 	 */
@@ -50,17 +46,17 @@ public class ParticleHelper {
 			{
 				for (int posZ = 0; posZ < factor; ++posZ)
 				{
-					double dirX = (double) x + ((double) posX + 0.5D) / (double) factor;
-					double dirY = (double) y + ((double) posY + 0.5D) / (double) factor;
-					double dirZ = (double) z + ((double) posZ + 0.5D) / (double) factor;
+					double dirX = x + (posX + 0.5D) / factor;
+					double dirY = y + (posY + 0.5D) / factor;
+					double dirZ = z + (posZ + 0.5D) / factor;
 
-					EntityDiggingFX particle = new EntityDiggingFX(world, dirX, dirY, dirZ, dirX - (double) x - 0.5D, dirY - (double) y - 0.5D, dirZ - (double) z - 0.5D, block, metadata);
+					EntityDiggingFX particle = new EntityDiggingFX(world, dirX, dirY, dirZ, dirX - x - 0.5D, dirY - y - 0.5D, dirZ - z - 0.5D, block, metadata);
 					effectRenderer.addEffect(particle.applyColourMultiplier(x, y, z));
 				}
 			}
 		}
 	}
-	
+
 	/**
 	 * Produces block hit particles at coordinates.
 	 */
@@ -69,7 +65,7 @@ public class ParticleHelper {
 		EntityDiggingFX particle = new EntityDiggingFX(TE.worldObj, x, y, z, 0.0D, 0.0D, 0.0D, block, metadata);
 		effectRenderer.addEffect(particle.applyColourMultiplier(target.blockX, target.blockY, target.blockZ).multiplyVelocity(0.2F).multipleParticleScaleBy(0.6F));
 	}
-	
+
 	/**
 	 * Returns block with considerations for the overlay and side of block
 	 * being interacted with.
@@ -99,8 +95,8 @@ public class ParticleHelper {
 				break;
 			}
 		}
-		
+
 		return block;
 	}
-	
+
 }
